test,perf(dry-run): address PR review — parallelize cases, split test file, drop any
- proposal-dry-runner: run a change's input cases CONCURRENTLY via Promise.all
  (gemini) — each case is an I/O-bound sandboxed subprocess on the materialize
  critical path; concurrency stays bounded by MAX_INPUT_CASES and the sequential
  outer change loop, and Promise.all preserves outcome order.
- split proposal-materialize-api.test.ts (was 594 > 500-line cap, coderabbit):
  shared fixtures → materialize-fixtures.ts; the dry-run wiring suite →
  proposal-materialize-dryrun-wiring.test.ts. api.test.ts now 372 lines.
- drop `as any` from the dry-run test helper — type the status as DryRunStatus
  (coderabbit; repo no-any rule).
