+++ addons/adapter-server/cap-adapter-server/__tests__/proposal-materialize-api.test.ts
@@ -13,240 +13,18 @@
  */
 
 import { describe, expect, test } from "bun:test";
-import type {
-  CommandLayer,
-  ExecutionDryRunProvider,
-  OntologyRegistry,
-  ProposalChange,
-  ProposalDefinition,
-} from "@linchkit/core";
-import type { CodeGenerationProvider, QualityGateRunner } from "@linchkit/core/server";
+import type { CommandLayer, OntologyRegistry, ProposalDefinition } from "@linchkit/core";
+import type { CodeGenerationProvider } from "@linchkit/core/server";
 import { Elysia } from "elysia";
 import {
-  type MaterializeEngine,
   mountProposalMaterializeAPI,
   runProposalMaterialization,
 } from "../src/proposal-materialize-api";
+import { GOOD, makeEngine, makeProposal, makeProvider, PASS_GATE } from "./materialize-fixtures";
 
 const BASE = "http://local.test";
-const GOOD = "export const deduct_inventory = 1;";
-
-// ── Fixtures ─────────────────────────────────────────────────
-
-function makeProposal(overrides: Partial<ProposalDefinition> = {}): ProposalDefinition {
-  const now = new Date();
-  return {
-    id: "prop-abc12345",
-    title: "Add deduct_inventory action",
-    description: "When an order is approved, deduct inventory",
-    author: { type: "ai", id: "pattern-detector", name: "Pattern Detector" },
-    capability: "cap-demo",
-    changeType: "minor",
-    changes: [{ target: "action", operation: "create", name: "deduct_inventory" }],
-    impact: {
-      schemasAffected: [],
-      actionsAffected: ["deduct_inventory"],
-      rulesAffected: [],
-      dependentsAffected: [],
-      migrationRequired: false,
-    },
-    status: "draft",
-    createdAt: now,
-    updatedAt: now,
-    ...overrides,
-  } as ProposalDefinition;
-}
-
-/** Provider spy returning a fixed source (or throwing). Records call count. */
-function makeProvider(opts: { source?: string; throws?: boolean } = {}): {
-  provider: CodeGenerationProvider;
-  calls: { count: number };
-} {
-  const calls = { count: 0 };
-  const provider: CodeGenerationProvider = {
-    async generateCode() {
-      calls.count += 1;
-      if (opts.throws) throw new Error("model exploded");
-      return opts.source ?? GOOD;
-    },
-  };
-  return { provider, calls };
-}
-
-/** A gate that passes everything (no Bun transpiler dependency in unit tests). */
-const PASS_GATE: QualityGateRunner = { check: async () => [] };
-
-/** Engine fake that mirrors the real draft-only `updateProposal` contract. */
-function makeEngine(proposal: ProposalDefinition | undefined): {
-  engine: MaterializeEngine;
-  updates: Array<{ id: string; changes?: ProposalChange[] }>;
-} {
-  const updates: Array<{ id: string; changes?: ProposalChange[] }> = [];
-  let current = proposal;
-  const engine: MaterializeEngine = {
-    getProposal(id) {
-      if (!current || current.id !== id) throw new Error(`Proposal "${id}" not found`);
-      return current;
-    },
-    updateProposal(id, u) {
-      if (!current || current.id !== id) throw new Error(`Proposal "${id}" not found`);
-      if (current.status !== "draft") {
-        throw new Error(`Cannot update proposal "${id}": expected status "draft"`);
-      }
-      updates.push({ id, changes: u.changes });
-      current = { ...current, changes: u.changes ?? current.changes };
-      return current;
-    },
-  };
-  return { engine, updates };
-}
-
-/** Fake execution dry-run runner returning a fixed status (or throwing). Spies jobs. */
-function makeDryRunProvider(opts: { status?: string; throws?: boolean } = {}): {
-  provider: ExecutionDryRunProvider;
-  calls: { count: number; changeNames: string[] };
-} {
-  const calls = { count: 0, changeNames: [] as string[] };
-  const provider: ExecutionDryRunProvider = {
-    async dryRun(job) {
-      calls.count += 1;
-      calls.changeNames.push(job.changeName);
-      if (opts.throws) throw new Error("sandbox unavailable");
-      return {
-        changeName: job.changeName,
-        target: job.target,
-        // biome-ignore lint/suspicious/noExplicitAny: test injects an arbitrary status
-        status: (opts.status ?? "passed") as any,
-        inputCaseId: job.inputCaseId,
-      };
-    },
-  };
-  return { provider, calls };
-}
 
 // ── runProposalMaterialization (orchestrator) ────────────────
-
-describe("runProposalMaterialization — execution dry-run wiring (Spec 70 P3)", () => {
-  test("stamps dryRunStatus on the materialized change and persists it", async () => {
-    const { engine, updates } = makeEngine(makeProposal());
-    const { provider } = makeProvider({ source: GOOD });
-    const { provider: dryRunProvider, calls } = makeDryRunProvider({ status: "passed" });
-
-    const outcome = await runProposalMaterialization("prop-abc12345", {
-      engine,
-      provider,
-      qualityGate: PASS_GATE,
-      dryRunProvider,
-    });
-
-    expect(outcome.kind).toBe("ok");
-    expect(calls.count).toBe(1);
-    expect(calls.changeNames).toEqual(["deduct_inventory"]);
-    if (outcome.kind === "ok") {
-      expect(outcome.proposal.changes[0]?.dryRunStatus).toBe("passed");
-    }
-    // The durable signal is part of the SAME persisted update as the source.
-    expect(updates.at(-1)?.changes?.[0]?.dryRunStatus).toBe("passed");
-    expect(updates.at(-1)?.changes?.[0]?.materializationStatus).toBe("materialized");
-  });
-
-  test("a forbidden_side_effect dry-run stamps that status (warn-only, still ok)", async () => {
-    const { engine } = makeEngine(makeProposal());
-    const { provider } = makeProvider({ source: GOOD });
-    const { provider: dryRunProvider } = makeDryRunProvider({ status: "forbidden_side_effect" });
-
-    const outcome = await runProposalMaterialization("prop-abc12345", {
-      engine,
-      provider,
-      qualityGate: PASS_GATE,
-      dryRunProvider,
-    });
-
-    expect(outcome.kind).toBe("ok");
-    if (outcome.kind === "ok") {
-      expect(outcome.proposal.changes[0]?.dryRunStatus).toBe("forbidden_side_effect");
-    }
-  });
-
-  test("a THROWING dry-run runner never fails materialization (advisory)", async () => {
-    const { engine, updates } = makeEngine(makeProposal());
-    const { provider } = makeProvider({ source: GOOD });
-    const { provider: dryRunProvider } = makeDryRunProvider({ throws: true });
-
-    const outcome = await runProposalMaterialization("prop-abc12345", {
-      engine,
-      provider,
-      qualityGate: PASS_GATE,
-      dryRunProvider,
-    });
-
-    // Materialization still succeeds; the per-case throw is contained as infra_error.
-    expect(outcome.kind).toBe("ok");
-    if (outcome.kind === "ok") {
-      expect(outcome.proposal.changes[0]?.materializationStatus).toBe("materialized");
-      expect(outcome.proposal.changes[0]?.dryRunStatus).toBe("infra_error");
-    }
-    expect(updates.at(-1)?.changes?.[0]?.materializationStatus).toBe("materialized");
-  });
-
-  test("scoped retry dry-runs ONLY the scoped change; out-of-scope dryRunStatus preserved", async () => {
-    // A 2-change draft: `untouched` was already materialized + dry-run "passed" in a
-    // prior request; this request scopes a retry to `scoped`. The dry-run must follow
-    // the materialization scope so it never overwrites `untouched`'s prior signal with
-    // a fresh (here "threw") result — mirroring the scoped re-materialize contract.
-    const proposal = makeProposal({
-      changes: [
-        { target: "action", operation: "create", name: "scoped" },
-        {
-          target: "action",
-          operation: "create",
-          name: "untouched",
-          materializationStatus: "materialized",
-          generatedSource: GOOD,
-          dryRunStatus: "passed",
-        },
-      ],
-    });
-    const { engine } = makeEngine(proposal);
-    const { provider } = makeProvider({ source: GOOD });
-    const { provider: dryRunProvider, calls } = makeDryRunProvider({ status: "threw" });
-
-    const outcome = await runProposalMaterialization("prop-abc12345", {
-      engine,
-      provider,
-      qualityGate: PASS_GATE,
-      dryRunProvider,
-      changeNames: ["scoped"],
-    });
-
-    expect(outcome.kind).toBe("ok");
-    expect(calls.changeNames).toEqual(["scoped"]);
-    if (outcome.kind === "ok") {
-      const scoped = outcome.proposal.changes.find((c) => c.name === "scoped");
-      const untouched = outcome.proposal.changes.find((c) => c.name === "untouched");
-      expect(scoped?.dryRunStatus).toBe("threw");
-      // Preserved, NOT overwritten by the scoped retry's dry-run.
-      expect(untouched?.dryRunStatus).toBe("passed");
-    }
-  });
-
-  test("no dryRunProvider → no dry-run runs, dryRunStatus stays undefined (off by default)", async () => {
-    const { engine } = makeEngine(makeProposal());
-    const { provider } = makeProvider({ source: GOOD });
-
-    const outcome = await runProposalMaterialization("prop-abc12345", {
-      engine,
-      provider,
-      qualityGate: PASS_GATE,
-    });
-
-    expect(outcome.kind).toBe("ok");
-    if (outcome.kind === "ok") {
-      expect(outcome.proposal.changes[0]?.materializationStatus).toBe("materialized");
-      expect(outcome.proposal.changes[0]?.dryRunStatus).toBeUndefined();
-    }
-  });
-});
 
 describe("runProposalMaterialization — guards", () => {
   test("non-draft proposal → not_draft; provider NOT called, nothing written", async () => {
